refactor: combined the websocket client and server stream implementation
* Related #540

[ci skip]
